split ouf the data set types from data-interactive-types
The full data-interactive-types depend on the slate editor types. The data set portion of the types will be the most commonly used by plugins.
