+++ v3/src/data-interactive/handlers/case-handler.ts
@@ -2,7 +2,8 @@ import { createCasesNotification } from "../../models/data/data-set-notification
 import { ICaseCreation } from "../../models/data/data-set-types"
 import { toV2Id, toV3CaseId } from "../../utilities/codap-utils"
 import { registerDIHandler } from "../data-interactive-handler"
-import { DIFullCase, DIHandler, DIResources, DIValues } from "../data-interactive-types"
+import { DIHandler, DIResources, DIValues } from "../data-interactive-types"
+import { DIFullCase } from "../data-interactive-data-set-types"
 import { attrNamesToIds } from "../data-interactive-utils"
 import { updateCasesBy } from "./handler-functions"
 import { dataContextNotFoundResult } from "./di-results"
