+++ v3/src/data-interactive/handlers/case-search-handler.test.ts
@@ -1,5 +1,5 @@
 import { toV2Id } from "../../utilities/codap-utils"
-import { DIFullCase } from "../data-interactive-types"
+import { DIFullCase } from "../data-interactive-data-set-types"
 import { diCaseSearchHandler } from "./case-search-handler"
 import { setupTestDataset } from "./handler-test-utils"
