+++ v3/src/data-interactive/handlers/collection-handler.test.ts
@@ -3,7 +3,8 @@ import { appState } from "../../models/app-state"
 import { getSharedCaseMetadataFromDataset, getSharedDataSets } from "../../models/shared/shared-data-utils"
 import { ICodapV2CollectionV3 } from "../../v2/codap-v2-types"
 import { toV2Id, toV3CollectionId } from "../../utilities/codap-utils"
-import { DICollection, DIDeleteCollectionResult, DIValues } from "../data-interactive-types"
+import { DIValues } from "../data-interactive-types"
+import { DICollection, DIDeleteCollectionResult } from "../data-interactive-data-set-types"
 import { diCollectionHandler } from "./collection-handler"
 import { setupTestDataset } from "./handler-test-utils"
