+++ v3/src/data-interactive/handlers/data-context-handler.test.ts
@@ -6,7 +6,8 @@ import { gDataBroker } from "../../models/data/data-broker"
 import { getSharedModelManager } from "../../models/tiles/tile-environment"
 import { toV2Id } from "../../utilities/codap-utils"
 import { ICodapV2DataContext } from "../../v2/codap-v2-types"
-import { DIDataContext, DIUpdateDataContext, DIValues } from "../data-interactive-types"
+import { DIValues } from "../data-interactive-types"
+import { DIDataContext, DIUpdateDataContext } from "../data-interactive-data-set-types"
 import { diDataContextHandler } from "./data-context-handler"
 import { setupTestDataset } from "./handler-test-utils"
 import "../../components/web-view/web-view-registration"
