+++ v3/src/data-interactive/handlers/item-by-id-handler.test.ts
@@ -1,5 +1,5 @@
 import { toV2Id } from "../../utilities/codap-utils"
-import { DIFullCase, DIItem, DIUpdateItemResult } from "../data-interactive-types"
+import { DIFullCase, DIItem, DIUpdateItemResult } from "../data-interactive-data-set-types"
 import { setupTestDataset } from "./handler-test-utils"
 import { diItemByIDHandler } from "./item-by-id-handler"
