+++ v3/src/data-interactive/handlers/item-search-handler.test.ts
@@ -1,5 +1,5 @@
 import { toV2Id } from "../../utilities/codap-utils"
-import { DIFullCase } from "../data-interactive-types"
+import { DIFullCase } from "../data-interactive-data-set-types"
 import { setupTestDataset } from "./handler-test-utils"
 import { diItemSearchHandler } from "./item-search-handler"
